Fix #66
The error is raised when the Page parameter of the  redcap url is bogus or isn’t design when the control reaches that part of the code. As the page doesn’t actually exist there will no data fields in it. This will set the $fields variable to null(and array_keys() doesn’t allow null types. To resolve it I have added a isset() conditional which checks if a variable is null.